adapter: Fix views synchronizer logic
57872b449 fixed a couple of bugs present in the views synchronizer. As
part of that work, a new "cache" was added to the views synchronizer to
keep track of which queries in the query status cache have already had
their views synchronized with the server. However, this commit
also introduced a bug: instead of looking for views for the queries that
we *haven't* yet checked, the views synchronizer was looking for views
only for the queries for which we *already synchronized views*.

This commit fixes the issue by reversing the boolean logic.

Change-Id: Ic63d4deac400cdca23b3c2f5517d7209351af625
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7516
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI