dataflow: Correct return type for SUM in Postgres
We were returning what MySQL was expecting, so this is just doing what
Postgres expects when appropriate.

This patch makes `readyset_dataflow` aware of `Dialect` in a way it
wasn't before, which although limited to this operator constructor may
be undesirable; but the alternative of having a secondary function for
determining the type which `mir_to_flow` would have to call separately
was equally if not more unwieldy.

Addresses: REA-4283
Change-Id: I2e7532b038f0be0e05347bd3f2f861d996221a25