treewide: update mysql_common
Update mysql_common to 0.32.3 in order to get the new collation data
dictionary support.

As part of this update, we also need to downgrade the version of
sqlformatter and prometheus-parse as they use itertools v0.12.1 which
is currently incompatible with criterion.rs and cause cargo check to
fail. We should be able to updated them once again when a release with
[1] is published.

[1]: https://github.com/bheisler/criterion.rs/pull/743

Ref: REA-4382
Closes: #1258

Change-Id: I36614184b749c96c0046c88ca5e1c6a2d186eff6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7505
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>