mysql: Clear prepared statement schema cache
If we closed or deallocated a statement, and later prepared a different
statement with the same ID, we would reuse the previously cached
resultset schema when encoding results which could lead to errors. Be
consistent about clearing the schema cache by removing entries from it
when explicitly closed/deallocated and before adding new entries.

Addresses: REA-4283
Change-Id: Iaa6dcf2521246ced8f7c87547505743044c6b54c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7542
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>