replicators: Add DDL vertical tests for MySQL
This commit adds DDL vertical tests for MySQL. The tests are a copy
of the PostgreSQL tests, with the necessary changes to make them work
and remove the unnecessary build of custom ENUM types which are not
supported to MySQL.

This commit also fixes REA-4453 by removing the views which we can
query when a new column is added to MySQL table.

This is the initial version that will be used to replay failures.

Ref: REA-4453
Fixes: #1271

Change-Id: I7c09688821e412c7ddd475e9bc4d2238ca56bff8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7586
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>