nom-sql: Parse bit string literals as BitVec
We were doing extraneous padding every time we turn a `BitVec` into
bytes, resulting in incorrect predication on caches created with
parameters on `BIT` columns.

We also improve bitvec resultset type support in logictests and add a
test that previously failed on Postgres. (The equivalent MySQL test is
still failing due to REA-3381.)

Change-Id: I85fcf99449a14e9ddfc9e82020e08183cb552fd6