readyset-data: Fix MySQL REAL data type conversion
MySQL REAL datatype is stored as a 8-bytes floating point number [0].
Currently when lowering we are converting it to Float, which is a
4-bytes floating point number. This commit changes the conversion to
Double, which is a 8-bytes floating point number.

Fixed DDL Vertical to use Binary Protocol when querying the tables
as the text protocol always bring return as Bytes and they get compared
as Strings instead of Double.

[0]: https://dev.mysql.com/doc/refman/8.4/en/storage-requirements.html#data-types-storage-reqs-numeric

Ref: 4444
Fixes: #1270
Change-Id: I09b2813d5eeeee0598739a7f62424d617e106ec7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7614
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>