[wip] dataflow-expression: Convert bits to appropriate MySQL types
This only implements expression lowering; we still don't correctly key
on BIT column types or other issues related to REA-3381.

This requires two suspect decisions that merit discussion.

First, we leave bitstring literals annotated with `DfType::BitVector`
types, as we do for Postgres, even though MySQL doesn't actually have
such a type. We do this unless we are lowering a top-level literal which
should resolve to the default 'real' type of a binary string. Otherwise,
we assume that higher-level lowering passes will cast to the proper type
as needed: i.e. to a binary string if we compare to a string, and to an
integer if we compare to a numeric type or do arithmetic. I expect
further testing is required to have more confidence in this assumption.

Second, for this to work with `IN`, for which we generate multiple
equality expressions and OR/AND them together, we now lower each new
equality expression as a whole, instead of only lowering the left and
right hand side separately. This way, the left and right sides will
coerce to the appropriate type, which could be different for each
element on the right hand side of the `IN`. This has the side effect
that we re-lower the left hand side for every element. I don't know how
big of a deal this is, nor have I thought of an alternative yet.

Addresses: REA-3381
Change-Id: I4b439fa86e78e32720c67ab383f1bc987ef3de0e