dataflow-state: Full materialization persistence
Currently, fully materialized nodes store all their data in memory,
which has the potential to OOM the process, and then get into a crash
loop as each successive restart will attempt to rebuild that full mat
node in memory, and OOM again (and again and ...).

This CL allows fully materialized nodes to store their data on disk
rather than in memory. In lieu of a full-on "disk spilling" mechanism,
this CL takes a first step in "writing full mat nodes to disk" by
reusing our existing rocksdb infra.

This CL retains the existing behavior of rebuilding the entire fully
materialized node at each restart, which allows us to be more liberal
in durability of the rocksdb files (sstables and WAL): meaning, since
we know we'll always be rebuilding, we don't need to flush to preserve
across restarts and we can avoid fsyncs and even writing to the WAL
at all.

Iteration on the rocksdb configuration specific to fully materialized
nodes will be taken up in future patches.

Release-Note-Core: Add configurable option to store
  fully materialized node data to disk rather than
  in memory.
Change-Id: Ia2be0eebe8cd2b8701338486699bcfbe64e53973
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7808
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>