nom-sql: Don't enforce target table on FKs
During the resolve_schemas pass, we visit all the tables and columns
in the foreign key constraints. Currently we don't do anything with
FKs and enforce table/schema resolve imposes a limitation in not been
able to snapshot tables with FKs that reference tables that are not
currently replicated.

This change adds a new method to allow visiting FKs and their columns
without enforcing the schema resolution. In case the target table is
not been replicated and the FK is not provided with db.table notation,
we add a placeholder schema to the Relation.

Fixes: REA-4473
Fixes: #1289
Change-Id: I32be7c134d0d669f0a0628f980c4363f6ae24ce0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7634
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>