replicators: Enhance MySQL Snapshot.
Adjust the MySQL snapshot to use the new snapshot type. The snapshot
type checks if table has a primary key or unique key and uses it to
define batches to query the table, making it less intrusive, specially
for large tables. In case the table does not have a primary key or
unique key, the snapshot will do a full table scan.

Fixes: REA-4477
Fixes: #1303

Release-Note-Core: Enhance MySQL snapshot to use Primary Key or unique
   key when available. This makes the snapshot less intrusive than a
   `SELECT *` (Full Table Scan) for large tables.

Change-Id: Iafda6ea6c74888262a0eea8bc1e880a3214b068a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7641
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>