nom-sql: Handle more cases of hex literals (part 2)
Insufficient committing and testing of the previous patch resulted in
code that didn't actually work.  The two oversights were:  failing to
check for hex before integers and an incorrect attempt to handle
odd-length literals in Postgres, which was checking the length of the
returned bytes, not the length of the literal.

(Some of this code has been written so as to help support odd-length
literals in the future, but they don't currently work.  The first
obvious problem is that the return type here returns bytes, not bits.)

Fixes: REA-4456
Change-Id: I99715c9f5b7b8cbbf9c0a4a507c1f5ff8bdb2f0f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7652
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>