replicators: Fixed correctness of MySQL timestamp handling
MySQL Timestamps are stored as UTC, but the server returns them in the
local timezone. This commit fixes the handling of MySQL timestamps to
ensure that they are correctly converted to UTC during snapshot/
replication and converted to local time when read from the database.

This commit also adds a test to ensure that MySQL timestamps are
correctly handled during snapshot/replication by comparing the
timestamps with the upstream database.

This commits also adds timestamp support into DDL Veritical tests.

Fixes: REA-4469
Closes: #1279

Release-Note-Core: Fixed correctness of MySQL timestamp handling.
  Now MySQL timestamps are correctly converted to UTC during snapshot/
  replication and converted to local time when read from the database.
Change-Id: I9d50fb66a52c015de7b613d0d7e614767569075d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7661
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>