server: Properly shutdown rocksdb WAL thread
When a worker receives a message on the `shutdown_tx` channel, we now
plumb through the event to the the new `State::shut_down` function
on each node of the dataflow graph. This allows Readyset to properly
shut down the rocksdb WAL flushing thread when a domain is killed (due
to other error) or when the process is exiting.

This CL also raises the priority of listening for events on the
`shutdown_tx` within `Worker::run()`, by adding the `biased` tag
before that block in the `select!`. We do this in several places in
Readyset, most notable in the adapter.

Release-Note-Core: Now properly shutting down the rocksdb WAL at
  process exit.
Change-Id: Ib2d1f19fb6046b5db475d017f52528f729a4fd03
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7660
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>