dataflow-state: Capture memtable size in deep size
Currently we only capture the live sstable sizes as part of
`PersistentState::deep_size_of()`. This CL includes the memtable sizes
in order to give a more accurate value. This is especially important
if the table is small enough to have not been flushed to disk yet.

Release-Note-Core: More accurately report the size
  of a persistent node by including the size of
  open memtables.
Change-Id: I28a41126743866a795e33b29dc24ba9c4e77feac
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7667
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>