client: Remove only use of unstable rustc_attrs feature
We enabled an unstable compiler feature in order to use an attribute
that the compiler says "will never be stable" in order to compress
an Option<usize> down from 16 bytes to 8.  Trying to save eight
bytes is admirable, but the tradeoff that was made is not in the
direction we want to go.

See prior justification at:
https://gerrit.readyset.name/c/readyset/+/2284/21/noria/noria/src/
lib.rs#139

Change-Id: I465cdd56011cc2a5629c7e3ca898f5cca13716dd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7679
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>