worker: expose materialization FrontierStrategy
We already have the mechanics to modify, at start time, the
materialization frontier strategy, but currently have no way to set
any value. This CL adds the cli flag/env var necessary to pass a
non-default value in.

The value of having this is to allow upquery testing that doesn't get
hampered by having the resultant values in the cache all the
time. Thus, setting the cli value to map to
`FrontierStrategy::AllPartial` is useful for testing queries. This is
more of a development time optimization than anything a user or
operator would set for a production deployment.

I also removed the long-dead variant `FrontierStrategy::Match` as it
is certainly no longer needed.

Change-Id: I9a79f6d86c9c14570ad9f65103d4123d6c3fa707
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7699
Tested-by: Buildkite CI
Reviewed-by: Vassili Zarouba <vassili@readyset.io>