replicators: Add an option to skip CREATE PUBLICATION in postgres
Readyset contains a command line option for disabling setting up
the DDL event triggers (--disable-setup-ddl-replication). This is
useful when the database user is not a superuser and ttriggers are
created manually.

CREATE PUBLICATION also requires superuser privileges. This commit
adds a new command line option, --disable_create_publication for
skipping publication creation. If this option is passed, the
publication has to be created manually before the replication starts.

Change-Id: If25d880f8730efd778c7d77d81c49ace428a0eea