replicators: Add an option to skip CREATE PUBLICATION in postgres
Readyset contains a command line option for disabling setting up
the DDL event triggers (--disable-setup-ddl-replication). This is
useful when the database user is not a superuser and triggers are
created manually.

CREATE PUBLICATION also requires superuser privileges. This commit
adds a new command line option, --disable-create-publication for
skipping publication creation. If this option is passed, the
publication has to be created manually before the replication starts.

Release-Note-Core: New option --disable-create-publication for skipping
creating a publication. Used when readyset is not running
as a superuser and the publication has been created manually.

Change-Id: If25d880f8730efd778c7d77d81c49ace428a0eea
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7705
Tested-by: Buildkite CI
Reviewed-by: Vassili Zarouba <vassili@readyset.io>
Reviewed-by: Jason Brown <jason.b@readyset.io>