util: Remove step_trait unstable compiler feature
This was only used in some utility code to add 1 to a generic type, but
there exists stable technology to accomplish the act of addition.

Since this stuff appears to only be used with usizes, I went ahead and
added a T: Copy bound, which I think cleans things up slightly.

Change-Id: I454bd98713061e487a321d3162ebcd8ffedeb5af
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7727
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>