system-benchmarks: Replace not-perf for pprof
`not-perf` is used by `system-benchmarks` as an embeddable library
that can create flamegraphs. Unfortunately, we have an un maintatined
fork of an unmaintained project. This CL replaces `not-perf` with
`pprof-rs`, a rust implementation of the pprof project.

`pprof-rs` is an easily embeddable library, which required making a
few changes to the way the library is embedded. Previously, we had an
undocumented, janky IPC between the `workload` app (the actual
benchmark driver) and the child process running readyset. I have
expanded this into a semi-documented, less-janky IPC to support a
reasonably simple message passing scheme. This was necessary as
`pprof` needs to run within the process that is to be profiled, rather
than being external to it.

I also removed a few, unused config options that were only applicable
to `not-perf`.

Change-Id: If0041df2a92452be1a7cccdc34d67c6fe0a50633
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7733
Tested-by: Buildkite CI
Reviewed-by: Vassili Zarouba <vassili@readyset.io>