benchmarks: update `serde_yaml` depedency
The `benchmarks` crate currently uses version 0.8.26 of
`serde_yaml`. The version has a transitive dependency on `yaml_rust`,
which is unmaintained and logging warnings (not failures) in `cargo
deny`. This CL bumps the version of `serde_yaml` to the latest, which
no longer depends on the unmaintained `yaml_rust`.

However, `serde_yaml` is itself now unmaintained (as explicitly noted
by the maintainer). Instead of rewriting parts of the `benchmark`
crate to use (yet another) new dependency, this CL takes the simple
approach and just updates the `serde_yaml`.

Change-Id: I4b9ffe3081c7c1664bc6900bd2518f7dfdfeb572