Note experimental features in experimental-features.json
The experimental-features.json file is a machine-readable list of
experimental features in ReadySet.  Each feature is specified as a JSON
object in an array of JSON objects.

This kind of sucks to maintain by hand, but there aren't too many.  I
initially tried to create an experimental features ArgGroup via our
usage of the clap crate, but I couldn't find a way to iterate the
arguments in the group.

To automate this, we should probably just wrap our #[arg(...)]
declarations in something that also tracks that the argument is for an
experimental feature.

Addresses: REA-4536
Change-Id: I0644b0c94ac6c96d6a0d795969997503f71845c4