system-benchmarks: Optionally dump graphviz
When debugging the performance of a `system-benchmarks` run, it would
be handy to have a graphviz image of the dataflow graph. This CL adds
that as an optional, cli flag.

In order to output an actual image format, this CL introduces a new
module in `readyset-util`, which has a function to call the `dot`
application (which must be preinstalled on the local machine, which is
usually not a problem).

Change-Id: I1cd258455efc0993f7cdec706ae83908ad2d2228