dataflow-state: Add some missing cases in eviction path
While scanning the eviction code, I noticed a missing case in a
match statement.  Since that might not have been the only mistake,
an audit was conducted of the entire file, which revealed two
additional accidental omissions.  Clearly the code could use some
refactoring via macros to make it more dense and to make these
errors easier to see, but in light of possible near-term renovation,
I'm just going to make things more correct for now.

Change-Id: Ib999845211ca077aa66c49d89ca499376d060644
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7746
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>