readyset-server: Make eviction tick check a little less confusing
It's not necessary to allocate a whole new future each time through
the loop, and we can eliminate half of the confusion by removing
the ability to specify no eviction interval.  It's a little doubtful
that eviction should be disableable at all, but even if you disagree,
it's enough to have one way to disable it (the memory limit).  Two
ways is overkill.

Change-Id: If5dfa558d013f416817239e91cec61b451213476