docker: Remove deprecated version specifier from compose files
Version 2 of Docker Compose, which went GA in April of 2022, no longer
uses the version specifier in the compose file:
https://docs.docker.com/compose/migrate/

Version 1 of Docker Compose stopped receiving updates about a year ago
in July 2023.

As of Docker 25.05, a warning is now printed to the command-line on
`docker compose up` if the deprecated version specifier exists in the
compose file.  Remove the version specifier from our compose files to
silence the warning.

Change-Id: I8b95dffb8d461917290712c44263208bd7f65d9c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7773
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>
Reviewed-by: Ron Hough <ron@readyset.io>