query-logger: QueryLogger::run() refactor
Simplify `QueryLogger::run()` by moving the creation of a new
`QueryLogger` to a `new()` function, and handle events in a dedicated
struct method. (An upshot to this is moving that logic out of the
`select!` macro allows rustfmt to properly format that code).

This CL also corrects the use of the `_span` which is never
referenced (nor were log statements falling under that span) by using
an `instrument` attribute on the `run()` function.

Change-Id: I2c3af4cbd69aa4a6df6119af50a8e8354f487da9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7792
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>