mysql: Add end-to-end datatypes round trip test
This fills a testing gap: we don't round trip every datatype through
Readyset with an upstream for MySQL. Tests that do some amount of
round-tripping, such as the DDL vertical tests, aren't targeting
datatype coverage; nor do they test snapshotting (Readyset is already
set up with replication before it starts creating tables).

There is an analogous test for Postgres, but it also doesn't test
snapshotting, and could possibly be improved as far as verifying
protocol compatibility: this test directly constructs mysql client
library values, instead of relying on automatic conversions from native
Rust types.

This tests arbitrary values for a subset of types that are easy to
generate. We plan to add further datatype support.

Further improvements could include verifying the resultset schema
returned when preparing the select statements, but our current mismatch
on charset/collation means we have to special-case when to actually
check it, and that's a bit cumbersome. It would also be nice to unify
this test with the postgres one, but it remains to be seen whether that
is worthwhile.

Addresses: REA-4598
Change-Id: Ibc690bda81473fdf4ee005773d573f6bcf19c9dd