dataflow-state: Clean up tmp/working rocksdb files
If readyset crashes or exists abnormally, it is not guaranteed to
clean up any of the temp/working rocksdb files. This CL introduces an
optional check at startup where it looks for the persistence
`WORKING_DIR`, and cleans up any files remaining.

Release-Note-Core: Clean up working dir files on start up to ensure
  no disk space is leaked after a crash.
Change-Id: I60295ced02454685b236ef4fd6952bff8012035b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7790
Tested-by: Buildkite CI
Reviewed-by: Johnathan Davis <jcd@readyset.io>