query-logging: Correctly calculate adhoc QueryIds
When `verbose` mode is enabled for query-logging, each permutation of
the parameters to an adhoc query creates a different `query_id`,
thus massively inflating the number of metrics we capture and
report in our prometheus endpoint. This is because the place where we
currently calculate the query_id happens way before we rewrite the
`SelectStatement`.

This CL adds an additional calulation of the query_id, after a
rewrite, but does so on the QueryLogger thread, and not on the hot path.

Additionally, this CL fixes the problem of `show proxied queries` and
`show caches` where the `count` column is always zero (even with
`--query-log-mode` set to `verbose`).

Fixes: REA-4599
Change-Id: Idbe913a370e9b80cd8791076d6da162d1164dc3b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7794
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>