[wip] mysql-srv: Allow sending invalid datetimes over the wire
If an invalid date is inserted on the upstream with the help of `SET
sql_mode = ALLOW_INVALID_DATES;`, we currently die when trying to write
them out even when proxying to upstream. This allows us to proxy them
even if they are invalid, which should only happen in this circumstance.
(That is to say, I can' think of a reason we would see invalid dates in
these places *except* when the user is, presumably, expecting them due
to setting `sql_mode`.)

Change-Id: I416bb505c72370a93679d944e0786ecfb4737e1b