mysql-srv: Allow serializing invalid dates
When proxying to upstream using the binary protocol (i.e. PREPARE), we
deserialize and reserialize the resultset. Since the upstream may have
the `NO_ZERO_DATE` and `NO_ZERO_IN_DATE` flags for `sql_mode` disabled,
or `ALLOW_INVALID_DATE` enabled, we cannot rely on turning date values
into `chrono::NaiveDateTime` which rejects invalid dates. Simply fall
back to `mysql_common`'s serialization for any date we encounter.

There is a question of whether this could be avoided in the first place
by not deserializing the resultset received from the upstream in the
first place, but that would be a bigger lift and likely break some
things, including at least updating the `EXPLAIN LAST STATEMENT` status.

Change-Id: I1c072c47e415fee37547e60618d695b4038dbbf9