readyset-sql-passes: Resolve schemas for foreign keys.
This commit adjust current pass to also set the default schema for
foreign keys reference tables that do not have a schema set. If a fk
is created without a fqn in the format `schema.table`, depending if
the referenced table is not already processed by Readyset, the schema
will remain null. The problem happens in the second time Readyset
boots up and tries to recreate all tables. This time the reference
table will be present and the schema will be set to the default schema,
causing the table body to be different from the one we stored
initially, causing the table to be dropped and recreating, causing a
new snapshot of the table.

Fixes: REA-4601

Release-Note-Core: Fixed a bug which caused tables with foreign keys
  to be recreated during the second start of Readyset when the reference
  table schema was not set.

Change-Id: Ia365b5003b037e8f5b3e7556ae4eff1849469068