replicators: Disallow DML events from STATEMENT binlog format
This commit adds a check to act on receiving a DML event via
STATEMENT binlog format. If we receive INSERT/DELETE/UPDATE events
we parse the event, extract the table, remove the base node from that
table and add the table as non-replicated.

Fixes: REA-4318
Fixes: #1211

Release-Note-Core: Added a functionality to abort replication of
  MySQL tables that receive DML events in STATEMENT binlog format.

Change-Id: I5fac7e0361a1957289e67dca3810fee93640ae36
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7812
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI