replicators: Refactor a couple of patterns for stabilization
The min_exhaustive_pattern feature has stabilized in current nightly (as
of and just prior to nightly-2024-08-12), but that compiler now
complains about irrefutable patterns.  We refactor the code slightly
so that it is accepted by both our current toolchain and current nightly
(although we are not switching toolchains at this time).

We have to add feature gates to get our current toolchain to accept the
refactored code, which seems opposite to the direction we want to go,
but all remaining feature gates will go away before long.

Change-Id: I9bdf3e0158a0d3e438766c74b67fbcae9df63a0f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7814
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>