readyset-server: Refactor Replica::run loop
Because rustfmt doesn't go into macros, it's a good idea to minimize the
code contained in them.  The new loop now fits onto my screen.

I moved around some log span entrances, but since in one place we
entered the span twice, it wasn't clear that there was great logic
behind the way it was before.

Change-Id: I4678bfa5dda848646e85acfc07bfaffe88d0ada0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7855
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>