mysql: Improve zero date support
We can now store and retrieve zero date values in caches. This does not
necessarily improve the situation for non-zero invalid dates (like
February 29th on non-leap years), which we convert to zero dates when
snapshotting or replicating them. There is nothing stopping us from
creating a cache keyed by a datetime and looking up invalid dates,
except that this is will currently only work with the binary protocol
(i.e. prepared statements). Another conversion is needed for coercing
text values to datetime for cache lookups to work with the text
protocol, and that requires more deliberating and testing.

Change-Id: I8729f8711ab673f6dc3642405730101400b65a69
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7815
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>