mysql-srv: Give precedence to long data params over NULL bitmap
Some clients, such as PHP's mysqli, will set the NULL bitmap bit for
params for which they actually sent long data blobs. MySQL appears to
give precedence to the long data, but we were giving precedence to the
null bitmap.

The new test is only a regression test that long data still works; it
would be cumbersome and fragile to test the new null bitmap behavior in
Rust land, since the mysql client library doesn't expose
`send_long_data`, nor does there seem to be a way to get it to set the
null bitmap bit for params which it sends via long data packets; we
would have to construct all the packets manually.

Fixes: REA-3162
Change-Id: Ib9513c95931e8ce88007ecae306a5e95bd1f569d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7822
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>