dataflow-expression: DfValue conversion error when case is used
- Fixes improper CASE statement result type inference.
The RC was that the result type was taken as the 1st THEN arm expression,
instead of inferring the "largest" common type over all THEN arms and
the ELSE expression. Error message should be returned if such type
can not be inferred. Note, a degenerated case when NULL only values are
present in all THEN and ELSE expressions, is considered as unsupported.

Fixes: REA-4442
Change-Id: I91e372da857663faee2e4380a8e9793584f2c4bd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7824
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>