readyset: Split database user into cdc and upstream
This commit splits the database user into two separate users: one for
upstream queries (proxy) and one for CDC (snapshot and replication).
This is necessary because the CDC user needs an elevated privilege to
read binlog/WAL and also, in case of postgres, to install the
triggers necessary for replicating DDLs.

This commit adds a new `--cdc-db-url` parameter to allow users to
specify a separate database URL for the CDC user. If not specified,
the CDC user will use the same URL as the upstream user.

Fixes: REA-4236
Fixes: #1282

Release-Note-Core: Added a new `--cdc-db-url` parameter to allow users
  to specify a separate database URL for the CDC user(used for
  snapshot and replication). This user can be a different user than
  the upstream user (used to proxy application queries). This allows
  users to have a specific user for replication with required
  privileges, without having to give extra privileges to the
  application user.

Change-Id: I0fa2e3d831cd12be6e261f96378afc942846386f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7843
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>