readyset-client: Reject some erroneous mixed type lookups
When preparing to query our cache, we wind up coercing the query
parameter literals to the types that exist in the cache.  This is a bit
different from the behavior of MySQL/PostgreSQL, which perform the
comparison after deciding on a sensible type promotion.

Previously, we were too permissive with our coercions for cache lookup.
If using a double to query a cache of integers, we would round the
double to the nearest integer and then use that value to perform the
lookup.  This behavior could produce erroneous cache hits.

Now, for known problematic comparisons, we only move forward with the
type coercion for a cache lookup if the original value can be accurately
represented by the type in the cache.  For example, if using a double
literal to perform a lookup on a cache of integers, we will reject 42.1,
but will accept 42.0 as a valid candidate for lookup.

While this prevents some erroneous cache hits, a more robust solution
would probably entail making our cache able to perform lookups using the
original literal value and type, performing type promotions for
comparisons as appropriate.

Also add some corresponding logictests.

Fixes: REA-4435
Change-Id: Ie888fc252fff4942cabcc1016b9ae2795b2cd0ae