readyset-server: Top level barrier manager for eviction
We introduce into the worker a barrier manager to ensure that each
eviction completely finishes before the next one starts.  Once
created, barriers are sent to downstream computation and can be
forwarded to an arbitrary amount of further downstream computation
through an arbitrary number of communication streams, all of which
is unknown to the barrier creator.  Only when all of the spawned
computation completes does the barrier complete, allowing the barrier
waiter to proceed.

The state of a barrier and its progress are tracked by means of credits.
A barrier begins with u128::MAX credits, which are split amongst the
downstream computations to which it is sent.  Then, the barrier manager
waits for the credits to come back.

Downstream computation may split received credits into an arbitrary
number of pieces, but because the barrier manager waits for all credits
to be returned, it does not have to know anything about what splitting
may or may not happen downstream.

The datatype used to represent the credits constrains the maximum
possible degree of splitting, which depends on the dataflow graph's
structure, which depends on user input.  We choose u128 as probably
sufficiently overkill so as to not underflow yet not so large as to be
unnecessarily wasteful.

In a future patch, domains will handle the onward marshaling of received
barrier credits.

Apart from the unit test that shows the operational principle, none of
this code is executed yet.  The dead_code turds are temporary only.

Change-Id: I1a79da8fd04e6e28d424f39017b69b8f4fc3dce1